+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-Robofirm DevOps Tool: MySQL Database Support
+Conductor: MySQL Database Support
 ============================================
 
 # 0.2.0 (Unreleased)
+- Renamed to Conductor
 - Updated PHP version requirement to 7.1
 - Cleaned up DatabaseAdapter code
 - Removed unused FilesystemTransfer classes
@@ -11,7 +12,7 @@ Robofirm DevOps Tool: MySQL Database Support
   since each operation is sufficiently complex to warrant a dedicated class
 
 # 0.1.1
-- Removed DI config for \DevopsToolCore\Database\DatabaseMetadataProviderInterface
+- Removed DI config for \ConductorCore\Database\DatabaseMetadataProviderInterface
 - Added DatabaseAdapterFactory
 - Merged DatabaseMetadataProvider into DatabaseAdapter
 - Fixed quoating of db name in dropTableIfExists and added exception on error